Removed Box of Hugs and Donk Pockets from Bounties (#26481)
As far as I know, the Box of Hugs only spawns on the clown and is used in the construction of a honkbot such that if the clown decided earlier to craft a honkbot or if the low price of some 3000 spesos is insufficient this bounty is impossible. Similarly, Donk Pockets are not a renewable resource and it has happened that a station has eaten too many for the bounty to be completed only to get it in rotation and esssentially lose a bounty slot forever. I would like it if each bounty was at least theoretically possible on any station it's likely to occur in, and there are too many where there aren't enough of these obtainable to complete the bounty, which is no fun.

(cherry picked from commit 3e786f548d2b429d052fb5c0d54b81c5234c9372)